Tried to minimize warnings from third party libraries.
git-svn-id: http://localhost:9111/private/safir_open/safir_sdk_core/trunk@334 9dd7cf02-b8c9-4ff8-a4a5-e7fda3bdd973

+++ src/dose/dose_communication.ss/src/dose_com/DoseOsInterface.cpp
@@ -318,7 +318,7 @@ unsigned long DoseOs::CreateThread(unsigned long& tid,
 {
 #ifdef _WIN32
 
-    tid = _beginthreadex(NULL,0,thread_function, param,0,NULL);
+    tid = static_cast<unsigned long>(_beginthreadex(NULL,0,thread_function, param,0,NULL));
 
     if (tid != 0x0L)
         return 0; // Ok
